requirements.txt
streamlit>=1.30.0
pydantic>=1.10.0
python-docx>=0.8.11
PyPDF2>=3.0.0
lxml>=4.9.0
openpyxl>=3.1.0
requests>=2.31.0
orjson>=3.9.0
PyYAML>=6.0.0
anthropic>=0.18.0
openai>=1.0.0
